Happy flow now working, alongisde UI animations

+++ app/src/main/java/com/drss/ifoodmobiletest/data/Sentiment.kt
@@ -1,3 +1,3 @@
 package com.drss.ifoodmobiletest.data
 
-data class Sentiment(val magnitude: Float, val score: Int)
+data class Sentiment(val magnitude: Float, val score: Float)
